[feat]: 피드 네비게이션 FeedId 타입 수정 (#96)

+++ app/src/main/java/com/texthip/thip/ui/navigator/extensions/FeedNavigationExtensions.kt
@@ -20,7 +20,7 @@ fun NavHostController.navigateToFeedWrite() {
 }
 
 // 피드 댓글으로
-fun NavHostController.navigateToFeedComment(feedId: Int) {
+fun NavHostController.navigateToFeedComment(feedId: Long) {
     navigate(FeedRoutes.Comment(feedId))
 }
